+++ index.html
@@ -28,6 +28,9 @@
 	</form>
 </section>
 
+<script src="scripts/theme.js"></script>
 <script src="scripts/rendererMain.js"></script>
+<script src="scripts/npmPackageInstaller.js"></script>
+<script src="scripts/commandClick.js"></script>
 </body>
 </html>
