+++ package.json
@@ -15,7 +15,8 @@
     "babel-cli": "babel src/cli.js > bin/cli.js",
     "babel-index": "babel src/index.js > index.js",
     "babel-postinstall": "babel --watch src/postInstall.js --out-file postInstall.js",
-    "postinstall": "node postInstall.js"
+    "postinstall": "node postInstall.js",
+    "say-hi": "echo 'hello!'"
   },
   "author": "Sam Eaton",
   "license": "MIT",
