+++ scripts/packageJsonInterface.js
@@ -1,0 +1,8 @@
+// 'use strict';
+// const fs = require('fs');
+// const path = require('path');
+
+// function() {
+
+// }
+"use strict";
