+++ src/scripts/toArray.js
@@ -1,0 +1,7 @@
+module.exports = function(list) {
+  const arr = [];
+  for (let i = 0, ii = list.length; i < ii; i++) {
+    arr.push(list[i]);
+  }
+  return arr;
+};
